优化一点内存占用空间和性能 更好的兼容clang 版本升级到2.2.1

+++ tutorial/test_async_event.cpp
@@ -71,7 +71,7 @@ void test_wait_any()
 
 	go[&]() -> future_t<>
 	{
-		for (int i = 0; i < _countof(evts); ++i)
+		for (size_t i = 0; i < _countof(evts); ++i)
 		{
 			intptr_t idx = co_await event_t::wait_any(evts);
 			std::cout << "event " << idx << " signal!" << std::endl;
